feature/LoginForm: Refactoring
